make it look beter, create one component for all sort elements
