--- src/components/SortBy/PaginationButtons/PaginationButton.module.scss
@@ -1,7 +1,0 @@
-@import './../../../styles';
-
-.limit {
-  // display: block;
- display: inline-block;
- width: 100%;
-}
